Fixed Path ignoring previous walls issue
The problem was caused because the flags used by Dijkstra's were not being reset after each run. So when you ran the algorithm a second time, the shortest path calculation ran the same (Because walls weren't a part of the previous node set calculation) and all the visited nodes were the same (Because the walls were still flagged as visited), but some nodes that were now visitable were now also being visited (Because they were previously flagged as not visited, but now we can).